Refactor project based on agreed conventions #32
The conventions are documented in the README file. This means that some
files were moved and renamed. Also the imports are adjusted, such that
imports where only the type is used are marked as such. This helps
the TypeScript compiler to emit more efficient code.
