+++ README.md
@@ -70,6 +70,18 @@ For more information on using the Angular CLI, including detailed command refere
 
 Generally, we are orientating ourselves at the default Typescript naming conventions. We try to avoid abbreviations.
 
+#### File names and locations
+
+Generally files are grouped based on topic. For files used by multiple topics the file is put into the shared folder.
+
+For file names for services and components we use the Angular standard with `.service` or `.component` suffixes.
+
+Definitions of types and interfaces are located in the models folder.
+
+Configs are located in the `config` folder. For each config there should be a corresponding interface definition in the `models/configs` folder.
+
+Constants should be avoided if possible and should be part of configs if needed. If a constant is required that should not be considered configurable, they are located in the `constants` folder. To make the usage of the constants more explicit, the values are scoped in an object that is exported. The object should also be typed `as const`.
+
 ### Angular conventions
 
 We use private readonly properties with `inject` to inject a service rather than using a constructor where we would provide them as arguments.
@@ -80,7 +92,11 @@ Whenever possible, a Github issue should be referenced in both branchname and co
 
 - Branches: `[feature|hotfix]/[xxx]-[name-of-branch]`, where `xxx` refers to a Github issue number and the `name-of-branch`
   is a short summary of the feature/hotfix.`
-- Commits: `#[xxx] Your commit message`, , where `xxx` refers to a Github issue number
+- Commits: `Your commit message #[xxx]`, , where `xxx` refers to a Github issue number. Make sure to write meaningful commit messages (see [this blog post](https://cbea.ms/git-commit/) for details)
 
 Our githooks check for both the branch name and the commit message, but they will only output a warning if they don't
 match. This is because there are times when you _might_ want to deviate from these rules.
+
+### Testing
+
+Due to the difficult nature of UI-testing, components are not tested. However, this means that components should contain as little logic as possible. Services and ngrx-effects should be tested.
