+++ src/app/app.component.ts
@@ -1,8 +1,8 @@
 import {Component, inject} from '@angular/core';
 import {RouterOutlet} from '@angular/router';
 import {TranslocoModule, TranslocoService} from '@jsverse/transloco';
-import {type Language} from './shared/types/language.types';
 import {ParameterService} from './stac/service/parameter.service';
+import {type Language} from './shared/models/language';
 
 @Component({
   selector: 'app-root',
