+++ src/app/shared/configs/language.config.ts
@@ -1,0 +1,5 @@
+import {type LanguageConfig} from '../models/configs/language-config';
+
+export const languageConfig = {
+  defaultLanguage: 'de',
+} as const satisfies LanguageConfig;
