--- src/app/shared/constants/language.constants.ts
@@ -1,4 +1,0 @@
-import {type Language} from '../types/language.types';
-
-export const supportedLanguages = ['de', 'fr', 'it', 'en'] as const;
-export const defaultLanguage = 'de' satisfies Language;
