+++ src/app/shared/constants/route-param.constants.ts
@@ -1,6 +1,6 @@
 /**
  * Exposes route params that can be used to build routes and to access parameters in a unified way.
  */
-export class RouteParamConstants {
-  public static readonly ERROR = 'error';
-}
+export const routeParamConstants = {
+  ERROR: 'error',
+} as const;
