+++ src/app/shared/models/configs/language-config.ts
@@ -1,0 +1,5 @@
+import {type Language} from '../language';
+
+export interface LanguageConfig {
+  defaultLanguage: Language;
+}
