+++ src/app/shared/models/configs/stac-config.ts
