+++ src/app/shared/models/language.ts
@@ -1,0 +1,2 @@
+export const languages = ['de', 'fr', 'it', 'en'] as const;
+export type Language = (typeof languages)[number];
