--- src/app/shared/types/language.types.ts
@@ -1,3 +1,0 @@
-import {supportedLanguages} from '../constants/language.constants';
-
-export type Language = (typeof supportedLanguages)[number];
