+++ src/app/stac/service/parameter.service.spec.ts
@@ -1,6 +1,6 @@
 import {TestBed} from '@angular/core/testing';
-import {Parameter} from '../../shared/models/parameter.types';
 import {CsvParameter, ParameterService} from './parameter.service';
+import {type Parameter} from '../../shared/models/parameter';
 
 describe('ParameterService', () => {
   let service: ParameterService;
